일정관리(ScheduleManagement) Read 구현

+++ src/main/java/com/sparta/personalproject/dto/ScheduleManagementResponseDto.java
@@ -11,7 +11,7 @@ public class ScheduleManagementResponseDto {
 
     private int passward;
 
-    private int date;
+    private double date;
 
     private String title;
