+++ src/main/java/com/sparta/personalproject/dto/ScheduleRequestDto.java
@@ -10,7 +10,7 @@ public class ScheduleRequestDto {
 
     private int passward;
 
-    private int date;
+    private double date;
 
     private String title;
