+++ src/main/java/com/sparta/personalproject/entity/ScheduleManagement.java
@@ -14,7 +14,7 @@ public class ScheduleManagement {
 
     private int passward;
 
-    private int date;
+    private double date;
 
     private String title;
